Applied patch by Tadej Borovšak for fixing over percent progress.
git-svn-id: https://svn.code.sf.net/p/imagination/code@131 6247ab75-7297-438c-a817-b4e4d5db8f37

+++ src/support.c
@@ -265,7 +265,7 @@ GdkPixbuf *img_load_pixbuf_from_file(gchar *filename)
 	return thumb;
 }
 
-slide_struct *img_set_slide_info(gint duration, gdouble speed, void	(*render), gint combo_transition_type_index, gchar *filename)
+slide_struct *img_set_slide_info(gint duration, guint speed, void (*render), gint combo_transition_type_index, gchar *filename)
 {
 	slide_struct *slide_info;
 	GdkPixbufFormat *pixbuf_format;
