+++ src/support.h
@@ -58,5 +58,5 @@ void img_set_statusbar_message(img_window_struct *, gint);
 void img_load_available_transitions(img_window_struct *);
 void img_show_file_chooser(SexyIconEntry *, SexyIconEntryPosition, int, img_window_struct *);
 GdkPixbuf *img_load_pixbuf_from_file(gchar *);
-slide_struct *img_set_slide_info(gint , gdouble , void	(*), gint , gchar *);
+slide_struct *img_set_slide_info(gint , guint , void (*), gint , gchar *);
 #endif
